Arm: Update to make publish more intuitive
Created function to just set the current/encoder values
and then call the publish function.
The JSONify will happen within publish.

The functionality works directly to the venv.
Although there is an issue where the ESP crashes.
This is caused due to the ESP reading every message it sends.
This issue will be fixed in the next commit